Update post-fixed-asset-transactions-posting-layers.md
the source documents (SO, PO, Pending vendor invoice and free text invoice) are supporting "Current" posting layers only. In case of the user selects a fixed asset in free text invoice (or other source document) the standard system will filter on the  book with Current posting layer and fill in the book field with it, if there are no current layer the book field will remain blank (unfortunately the user could enter the non Current layer book name by keying it in since it is not a hard filtration the book is still there in the back end), we have implemented the the validation to check if the posting layers of the fixed asset in the document is not Current layer during posting to avoid data issues that could be generated. The recommendation is to use derived books to post to non Current layer for posted transactions from free text invoice.